Renamed costfunctions to functions. Added java doc.
