+++ simmetrics-core/src/main/java/org/simmetrics/metrics/functions/ConstantGap.java
@@ -20,62 +20,53 @@
  * SimMetrics. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.simmetrics.metrics.costfunctions;
+package org.simmetrics.metrics.functions;
 
 import static com.google.common.base.Preconditions.checkArgument;
 
 /**
- * A gap function that calculates the gap penalty as <code>A+(B * GapLegth)</code>.
+ * A gap function that assigns a constant penalty to a gap regardless of size.
  * 
  * @author mpkorstanje
- * 
- * @see Substitution
  * @see <a href="https://en.wikipedia.org/wiki/Gap_penalty">Wikipedia - Gap
  *      Penalty</a>
- *
  */
-public final class AffineGap implements Gap {
+public final class ConstantGap implements Gap {
 
-	private final float startValue;
 	private final float gapValue;
 
 	/**
 	 * Constructs a constant gap function that assigns a penalty of
-	 * <code>startValue + gapValue * gapLenght </code> to a gap. .
+	 * <code>gapValue</code> to a gap. .
 	 * 
-	 * @param startValue
-	 *            a non-positive initial penalty for creating a gap
 	 * @param gapValue
 	 *            a non-positive constant gap value
 	 */
-	public AffineGap(float startValue, float gapValue) {
-		checkArgument(startValue <= 0.0f);
+	public ConstantGap(float gapValue) {
 		checkArgument(gapValue <= 0.0f);
 
-		this.startValue = startValue;
 		this.gapValue = gapValue;
 	}
 
 	@Override
 	public final float value(int fromIndex, int toIndex) {
 		checkArgument(fromIndex < toIndex, "fromIndex must be before toIndex");
-		return startValue + gapValue * (toIndex - fromIndex - 1);
+		return gapValue;
 	}
 
 	@Override
 	public final float max() {
-		return startValue;
+		return gapValue;
 	}
 
 	@Override
 	public final float min() {
-		return Float.NEGATIVE_INFINITY;
+		return gapValue;
 	}
 
 	@Override
 	public String toString() {
-		return "AffineGap [startValue=" + startValue + ", gapValue=" + gapValue
-				+ "]";
+		return "ConstantGap [gapValue=" + gapValue + "]";
 	}
 
 }
