+++ simmetrics-core/src/test/java/org/simmetrics/metrics/costfunctions/SubstitutionTest.java
@@ -24,7 +24,7 @@ import static org.junit.Assert.*;
 
 import org.junit.Before;
 import org.junit.Test;
-import org.simmetrics.metrics.costfunctions.Substitution;
+import org.simmetrics.metrics.functions.Substitution;
 
 public abstract class SubstitutionTest {
